Avoid requiring encodeMatcher at launch if not needed
